Time-variable types changed from int to double

+++ sbnobj/Common/CRT/CRTHit.hh
@@ -27,11 +27,11 @@ namespace sbn::crt {
       float         peshit; ///< Total photo-electron (PE) in a crt hit.
      
       uint64_t       ts0_s; ///< Second-only part of timestamp T0.
-      int8_t    ts0_s_corr; ///< [Honestly, not sure at this point, it was there since long time (BB)]
+      double    ts0_s_corr; ///< [Honestly, not sure at this point, it was there since long time (BB)]
       
-      uint64_t      ts0_ns; ///< Timestamp T0 (from White Rabbit), in UTC absolute time scale in nanoseconds from the Epoch.
-      int8_t   ts0_ns_corr; ///< [Honestly, not sure at this point, it was there since long time (BB)]
-      int64_t       ts1_ns; ///< Timestamp T1 ([signal time w.r.t. Trigger time]), in UTC absolute time scale in nanoseconds from the Epoch.
+      double        ts0_ns; ///< Timestamp T0 (from White Rabbit), in UTC absolute time scale in nanoseconds from the Epoch.
+      double   ts0_ns_corr; ///< [Honestly, not sure at this point, it was there since long time (BB)]
+      double        ts1_ns; ///< Timestamp T1 ([signal time w.r.t. Trigger time]), in UTC absolute time scale in nanoseconds from the Epoch.
      
       int            plane; ///< Name of the CRT wall (in the form of numbers).
 
@@ -46,8 +46,8 @@ namespace sbn::crt {
 
       CRTHit() {}
 
-      int64_t ts0() const { return static_cast<int64_t>(ts0_s) * 1'000'000'000LL + ts0_ns; }
-      int64_t ts1() const { return static_cast<int64_t>(ts0_s) * 1'000'000'000LL + ts1_ns; }
+      int64_t ts0() const { return static_cast<int64_t>(ts0_s) * 1'000'000'000LL + static_cast<int64_t>(ts0_ns); }
+      int64_t ts1() const { return static_cast<int64_t>(ts0_s) * 1'000'000'000LL + static_cast<int64_t>(ts1_ns); }
 
     };
